requirements.txt
requests
slack_bolt
atlassian-python-api